collector/requirements.txt
paho-mqtt==2.1.0
msgpack==1.0.8
PyYAML==6.0.2
